* Added AddMenu, SetMenu, DrawMenuBar, CelWide, CelHigh, AddToPic * Fixed memory management system according to Lars' specs * Fixed NewWindow to redraw with ylo-11 if (flags & 0x4) (if it has a title), and set the global port to the new window. * Added a menu management system. Menus are copied from the script during AddMenu() time, and the menu entries are split up. This should make menu box drawing much easier later on. * The Fun Seeker's Guide now displays correctly (exclamation mark), with the following exceptions (you probably expected this...):   - The information text doesn't display (kfunct Display())   - The Prev and Next arrows don't display (kfunct DrawControl())
DrawControl() has already been documented by Lars, but Display() looks funny.
It takes quite a bunch of arguments, some of which refer to an address
relative to script.1 (script.1, not text.1).
BaseSetter() also appears to do funny things (apparently it sets brLeft,
brRight etc. to form a bounding box around the coordinate varselectors.
Don't know what a bounding box around a single pixel is supposed to be good
for, though).
Also, I'm not sure how Animate() work. Fortunately graphics seem to be the
next interesting thing on Lars' list of Things To Document.


And a note to all those lurkers on the list: I'd appreciate it if someone
would volunteer to take care of the docs. I haven't updated them in
a week or so, and we've got several new things since then.

This will probably be my last CVS update for this week, as RL is approaching
(and doesn't appear to be in a good mood).

BTW: Does /anyone/ read those CVS logs, anyway?

 - Christoph