* Implemented StrEnd, StrCat, StrCmp, StrCpy, StrAt, ReadNumber, DrawControl * Fixed Format to allow the usage of strings from the heap * With DrawControl in place (though not fully tested yet), the FSG displays exactly as in the DOS executable (Well, I haven't checked every pixel). * Improved disassembler to dump kernel function parameters and the name of the object/class/clone for send and family.
Things look good so far. Meanwhile, I think that the problems with the SQ3
animation not working simply come from the Bresenham stuff being unimplemented,
so it's not a bug.

BTW: We've got 58 "theoretically complete" kernel functions now; that's more
than half of the total 114/115 SCI0 kernel functions (and many of those
ahead aren't really needed for FreeSCI, like FlushResources() or Profile()).

 - Christoph