* Fixed a bug in resource unloading that would cause infinite recurstion (I was pretty sure that the thing was designed not to recurse infinitely, but I was the only one who played with it...) * Fixed the selector drawing bug (actually a bug in Format()) that caused LSL2 to die horribly. ATM it looks as if windows in QfG1 and LSL2 default to a different alignment (QfG1 uses centered, LSL2 uses left-aligned); this might be the first version difference. MIGHT be- QfG1 uses DrawSelector for drawing its text, but I'm not so sure about LSL2.  Anyway, I'll add a "versions.h" and some versions checking macros tomorrow... I mean later today.
For people with CVS access, this version might be worthy to be checked
against one or more of the games I havent' checked it against. If nothing
unexpected pops up, I'll do a new release later this week; probably friday or
saturday. This means that the docs will have to be updated. Hm.

 - Christoph