* Added NumCels and NumLoops. * Fixed Animate() after I found out about the nasty things it does with the "doit" function selector when nobody's watching. Of course my implementation is completely wrong; fortunately Lars unravelled the mysteries of the "signal" selector for us, so this isn't going to be fatal. * Fixed countless bugs ranging from drawView0() not inversing the relative x offset if the loop is reversed over numerous sign extension bugs to bugs deep inside script unloading (script unloading has been made MUCH simpler). * Version checking is in place now (have a look at src/include/versions.h; version checks can be done by directly comparing version numbers, since sci_version_t is typecast to be an int). Of course, in order for versions checking to be effective, some way of determining the version number needed to be added. The obvious way would have been to pass it on the command line... * Added a config file system (lex sources will be part of the distribution). The config file is called either ~/.freesci/config or ./freesci.cfg (for people without a home like DOS users). The ~/.freesci directory also contains subdirs for each game you have played/will play; these subdirs will hold savegames later on. The config file format is rather simple; here's an example: ------------------------------------- # This is a comment version = 0.000.685 # This sets the default version for all games to 0.000.685. # Incidently, 0.000.685 also is the default if none is specified.
color_mode = dither
# Set 16 color dithering. Other options would be "interpolate" or "dither256"

[SQ3]
# The name is retrieved from the game object

version = 0.000.435
# This will only affect SQ3
-------------------------------------
This needs to be documented anyway.

Well, just for fun, here are the results of the latest tests:
+ LSL2: Does everything right, up to and including the input box. Text
is centered correctly if the version number is set up right.
(Maybe the change in text alignment was the reason for which they dropped
the "Enter Input" label in the standard input window)
+ LSL3: Animates somewhat, but Larry appears to steal Patti's right leg.
This is an unusual behaviour, even for Larry.
+ QfG1: Unchanged. Perfect. In a way. ;->
+ QfG2: Now looks like QfG1 (SCI01 isn't /that/ much different, after all),
except for the titlebar, which is drawn incorrectly.
+ KQ4: Draws the copy protection question box, and that's all I ask for. ATM.
+ SQ3: With the correct setup, my historical vintage-0.000.435 SQ3 does the
intro very nicely, up to the point where Scott Murphy and Mark Crowe show up.
(Maybe it's their fault). Even the paint on the "Pirates of Pestulon" drips
correctly. However, the Sierra logo is only displayed in the quick sequence
(no flying parts of the logo). This is because FreeSCI fails the speed check
(output to an x window is quite slow... honestly, it's not the engine ;-)


TODO:
 - UPDATE THE DOCUMENTATION. 10 Euros for anyone who comes over here and kicks
me until I do that.
 - Fix Animate() according to Lars' divine directions
 - Change picture_t. It needs to become a struct that also contains
information that allows other functions to use 16 bit graphics inside the
interpreter (we'll still need 8 bit graphics as well, for the DOS port).

 - Christoph