* Fixed numerous bugs * Added InitBresen, DoBresen, CanBeHere, DrawCel * Fixed some more bugs
Results:
- SQ3 runs extremely smooth now. Except for the Copyright message overdraw
problem, everything looks right. However, the game enters an infinite loop
(like LSL3) right before the escape pod animation is to be drawn. I assume
that it's waiting for a cue from the audio system, but I can't be sure.
- QG1, QG2, KQ4, LSL2, LSL3: Unchanged. QG2 no longer has the empty white
title bar during the legal notice, but that's it.


If it runs on the Alpha, this will be the 0.2.4 release.

 - Christoph