* Fixed Format(). * Fixed Animate(): Some elements of the display list add new entries to that list (while the kernel is working on it). * Did a DoSound() workaround: Each sound object called gets its lastPri selector set to -1 and state to 2. This appears to tell the interpreter that the sound object is finished.
Results:
+ SQ3 plays the complete intro. There are three bugs left to eliminate (two
of them might be related), but then it should work.
+ LSL3 continues with the intro, but has much worse problems.

Note:
- Win32 support is not in yet. Have a nap and try again tomorrow.

Solaris tests:
I tried to compile it under Sparc-Sun-Slowaris (a rather badly maintained
installation). It broke because:
- libreadline wasn't there
- nothing obstack-related was available
- getInt16() wasn't available (Chris made me stumble over that one yesterday).

So there was nothing really fatal. A cleanup of main.c needs to be done,
though; and the obstack in midi.c needs to be killed off. I may get another
chance to run a sparc test next week.

 - Christoph