* Implemented NULL sound server and generic sound server helper commands.   Portability-inhibiting commands used are:   + select()   + fork()   + pipe()   + socketpair()   I've tried to build it so that it won't break if no sound driver is   available. Dmitry, just shout if it breaks the Win32 port, and I'll   check it again :-) * Solidified NewWindow() * Fixed a bug in the unsigned compares where pprev wasn't set (Thanks, Lars!) * It turns out that the stack should NOT start before heap offset 1000, since   text commands use it to buffer their strings (D'oh! Should've thought of   that much earlier!). Fixed this in heap.c (this was the easiest way to   fix it). * Added debug mode 'a' for "audio"
Results:
+ QfG1 animates the char select screen correctly now. The cues are sent,
  and everybody's happy.
+ The mystery text box in the QG1 main screen is the "First Bug in this game"
  text box. Looks as if GlobalToLocal is broken.

Todo:
- The QG1 character setup screen is horribly broken. It's possible to get
  past it, but that's all.
- Walking doesn't work; I get a stack overflow each time I try (SQ3 and QG1).

So, the sound server appears to work. It doesn't play any sound, but nobody
ever said that a sound server should do that (Adding OSS support there
is a 0.3.x Issue (I should get a trademark for that). Win32 does things
differently, and Dmitry appears to have an idea about how to implement this,
so I'll keep my fingers out of it. How about ALSA? Have they agreed on a
MIDI interface yet? Or will ALSA users have to wait for sw synth support?)

Anyway, the sound server system should be fairly modular. The protocol
defined in sound.h /has/ to be used, since it's invoked by the kernel, but
the other stuff in soundserver.h is just there for convenience.


And, developers, please sign with your names if you check in something.
The mailing list forwarding script doesn't add a name, so no one will know
who to blame if something breaks ;-)

 - Christoph