* Fixed all console commands to get a state_t* parameter. * Brought back the on-screen command console. You knew this was coming   sooner or later ;-)   The new palette makes transparency a bit easier, and, IMHO, it looks better   now (using an EGA color bitmap instead of one color would be possible as   well- if someone makes one, I'll include it. Quake has one, too, after all ;)   The console still needs some cleanup. * Removed the "slog" command; introduced the "debug_flag" variable instead   (this way, all step commands can use logging) * Fixed displaying on 8 bit GGI displays. You can now display FreeSCI on   SVGAlib on any Intel box (using mode 0x13), if you are root. (Not re-   commended; use fbcon if possible) * Set debug mode to false as a default (This should become a command line   option).
The revenge of the on-screen command console means that people interested in
ports to non-windowing systems (DOS, PalmOS, Gameboy) can get back to work
now :-)

To prove that it's stable enough, I ran it on the SVGAlib target, as root.
The result was, of course, that it's not stable enough; when I quit, it
started eating up memory for no apparent reason (this obviously needs to
be investigated). That was one of those days where the magic SysReq keys
really came in handy.

 - Christoph