I'm not dead yet...
* Added experimental Parse() implementation based on Lars' specs.
* added "dumpnodes" command to examine the Parse() tree nodes.

This implementation produces the following structure for "look room" in SQ3:
<0141,<013f,<<0141,<013c,<<0141,<013b,<<0141,<013b,<<0141,<013b,<<0141,<013b,<<0000,0474>,NULL>>>,NULL>>>,NULL>>>,NULL>>>,NULL>>>,<0141,<013c,<<0141,<013b,<<0141,<013b,<<0141,<013b,<<0141,<013b,<<0000,080f>,NULL>>>,NULL>>>,NULL>>>,NULL>>>,NULL>>>>>>

I'm not convinced that it's even approximately correct, but maybe it does
something useful with the Said() trees that are supposed to come in the near
future. (And yes, the implementation is dirty.)

BTW: The cheat mode in SQ3 is activated by using "backstage pass" or
"pump shark". The Said() block unfrobber appears to work quite nicely when
it comes to figuring out this stuff.
(Available debug commands appear to be limited to "qa" and "tp").

I'm very busy at work ATM. Expect updates from my part to come in small and
rare droplets during the next month or so.

 - Christoph