In vm.c and, script.c:
Improved support for the old SCI versions. Sorry if I broke something.

In main.c:

Fixed a bug in the configuration system; if the emulated version was changed by game_init_state (as in the case of old SCI's), the default configuration setting would override this. Don't know if I did this in the best possible way.

       Lars