* Fixed the "out of hunk handles" problem. I hope. Apparently, views that   are about to be disposed expect the kernel to handle their underBits   automatically. * More BaseSetter improvements; BaseSetter now takes the view offset into   consideration. * Format ("%c.", 'a') now outputs "a." instead of 'a%'. * Patches for NetBSD according to Claudio's directions * Made sure that &rest is stored in save games. This only affects the   console save commands; the kernel SaveGame() always has its &rest set   to zero, since it's callk'd. Well, it would, if it was implemented. * Ripped apart acinclude.m4. It doesn't check for files any longer; instead,   it tries to run the preprocessor (for header files) or a complete compile/   link cycle (for libraries). Defun'd AC_CHECK_INCLUDE_PATH and   AC_CHECK_LINK_PATH who try preprocessing/linking with a list of options,   respectively.   They always try to link without appendices first, though.   Using --with-xxx-yyy will enforce that setting now; this will be the only   possibility tested (except for "").   It should now be possible to specify whether linking should go against   curses or ncurses.
Please check the new configure script. It should work better (especially
for cross-compiling), but I haven't tested it on any box other than my own yet.

As for the release: I'd like to do that one asap, but the execution stack
problem must be taken care of first, and the configure script needs to be
tested. Of course, development releases aren't really that important if
CVS snapshots are available, but they're a good excuse for laying off new
features and concentrating on fixes once in a while :-)
I'll give people time to complain about the new configure script until friday;
with some luck, we'll have a few more bug fixes by then.

-- Christoph