* Added code to check for unaligned (16 bit) access. Preliminary tests   indicate that those are either very rare or don't happen at all.   In theory, this would allow for optimizations for big-endian systems;   unfortunately, instruction fetch isn't considered there (and every   second fetch is unaligned). * Fixed a bug in the title bar: When drawing the status bar, the newline   character is not interpreted as a special character. This fixes the   KQ4 title bar bug (which nobody has observed yet because the next fix   hadn't been applied :-) * Added preliminary restart support. The first restart calls play(), which   KQ4 (and probably PQ2) use for starting the actual game, the second and   following ones call replay(), which are used for the "real" restart.   I guess that the games use a different mechanism to indicate this, though.   Note that the play() restart appears to be clean, while the replay() one   leaves the stack in a semi-messy state. * Played around with button and input box positioning. Might be a bit better   now. * Added the 'meminfo' command. Useful for triggering failed assertions if   the heap is corrupted.
It's been quite a while since the last update. Anyway, I'm Not Dead Yet,
and this project will continue, even if the pace is a bit slow right now.
We're really close to something very useable- almost all SCI0 features are
in (savegame menus and the text parser being the most notable exceptions,
not counting 0.3.x issues like sound), and the number of entries in the
list of open bugs can still be described with one digit (using base36
notation).

BTW, owners of KQ4 might want to check out that game now. The images in the
intro are still screwed, but it's finally possible to play the real game
(you can even skip the copy protection with the 'restart' console command,
but then you won't be able to enjoy all the interesting bugs in the intro).

-- Christoph