* Fixed a problem with X11 header file locations that didn't require a special   -L directive * Fixed drawing of BG images with signal 0x8000 ('dispose me') set. They're   drawn before all other images read their BG buffer, thus fixing most visual   glitches in the KQ4 intro. * Added Rickard to the list of people mentioned in the startup message. * Fixed a problem introduced by the new startup code, which would cause all   scripts required by script 0 to be loaded and then to be set inactive,   without freeing the memory they took on the heap. * Fixed a sound server bug that would cause a deadlock in some situations   where the end of a song was reached. * Fixed a few sign extension bugs, most notably in &rest.
As hinted above, the KQ4 intro is played correctly now, except for one
priority bug (framed flying adventurer's hat). Actually, KQ4 provides an
interesting test bed now, since it allows a lot of area to be explored
without requiring any commands to be typed.

BTW, I'd appreciate information about any of the following:
- SetNowSeen()
- EmptyList()
- AddAfter()
- DoAvoider()
- TimesTan()
- TimesCot()
- DisposeList() when called with a node (as opposed to a list) as parameter
- Said()


My pay job's officially over for the time being, but lectures are starting
again next week; I'll try to reserve some time for FreeSCI for the next few
weeks, though. I'd like to do a new release in a month or so; it'd be nice
if this could be the 0.3.0 release, but I'm willing to settle for 0.2.6.

TODO:
- Investigate the functions mentioned above (especially Said())
- Add missing Graph() functionality
- Fix savegames
- Fix list selection widget (the thing used to select savegames)
- Test restarts, menus, other generic stuff
- Update documentation

-- Christoph