* Added -Wall to the compiler options, removed a few warnings * Fixed bug in the suffix vocabulary * Fixed problem in the class detection algorithm (>= instead of <= was added   during some code beautification) * Fixed Said() some more * Added parse tree augmenting
Observations:
+ A few... more warnings. I knew I had had a reason for not using -Wall ;-)
+ Technically, the text parser is working, and I think it's quite close to
  how the original worked. There are a few problems with this, though:
- Throwing a rock in QfG always gets that rock thrown to the upper left of
  the screen (0,0).
- Trying to look at inventory items in QfG fails with invalid Said specs
  (broken addresses)

Both problems might be related to a problem with local variable reading or
writing (or the related (a|p|s)To(p|s|a) operations).

(Note that menubar Said() specs aren't supported yet. Adding this should
be trivial, though).

Have fun experimenting with this. It's theoretically possible to finish all
supported games now.

-- Christoph