* Updated metadata
The *NIX port is ready for the release.
Systems this release has been tested on:
- i686-pc-linux-gnu
- sparc-sun-solaris2.5
- sgi-mips-irix6.2 (Indy R5000 using native cc, experimental)

I'm waiting for the ports, though. I still don't know whether the win32 port
works, but the DOS one appears to be worked on.
Anyway, 0.3.0 will be released when the ports are ready (or when it's obvious
that they're dead). I'll wait a few days; if it takes too long, we'll just
split the tree into a 0.3.0 and a 0.3.1 branch for experimenting with sound,
graphics, and related stuff.

-- Christoph