+ Fixed bugs in the dynviews implementation + Fixed mirrored cels + Implemented display lists (for the most part)
With the global display list back, there's finally something to see again,
so I took the liberty of making two screenshots and posting them to our
homepage to demonstrate that this project isn't quite dead yet.

there are plenty of bugs left in it, though, and I'll try to put them down
or fix them until the end of the week. I guess it's time the debugger is
fixed.

-- Christoph