* More portability fixes as suggested by Petr * The libggi API recently changed the semantics of ggi_mode.size to mean   mm instead of pixels- used ggi_mode.visible instead. * Implemented on_control style functions (not fully correct yet). * Enabled the standard debugger (not the on-screen one) again. * Implemented three gfx_* debug functions
The libggi API change problem caused an FPE in what appeared to be a
corrupted stack. However, this was caused by a division by zero in the
initialization part of a function that had just been linked dynamically,
which, in turn, was caused by ggi_mode.size.(x,y)/(320,200) returning
zero. Figuring this one out was... interesting.

-- Christoph