src/engine/ksound.h: Implemented _K_SOUND_VOLUME src/include/engine.h: Added volume/mute to game state src/sound/sound.c:  Implemented VOLUME and MUTE commands src/sound/soundserver_null.c:  Debug output on VOLUME & MUTE
Basically, the first round of sound changes.  Now the kernel and
internals should know everything necessary for sound stuffs.

Next step is to up the ante and make the null sound server start to
print out midi events.  I want to make sure it's reasonably correct
before I work on the hardware side of things.  :)

Questions:

Where would be the best place to initialize the new state volume/mute
variables?  I presume main.c's init_gamestate() is the place to do it,
but ideally, there should be a sound_volume setting in the config file,
(and that stuff should have its own initializer functions) and if
that's not present, it should default to 0xc.    Yes, the sound
volume is unititialized until you set it in the game.  :)  (I'll fix
this as soon as I figure out the config stuffs)

Apparently when you set the volume, it performs at least two actual sets
-- it sets it to 0, then sets it to whatever value you want.  I don't
think that's my code, but rather what the interpreter is doing..

Wheee!  I think I'll go to bed now.  My laptop battery is about to die,
so I don't have much choice..

 - Pizza