Began some of the work needed to merge the new sound output system.
Moved midi.c to oldmidi.c
changed all of the headers to use glib.h (is this really necessary?)

Next up:  Add configuration stuffs to the config file.  The idea is to
have this thing (at the very least) dumping the PATCH.001 file on
initialization, based on the config file.

 - Pizza