* Various build fixes for 'make distcheck' * sciv.6 man page updated for 0.3.1 * Added keysym -> scancode << 8 converter for cases where ALT-something   is pressed
That third fix enables some cheats within a limited number of games and
should not have any major impact.

Observations:
+ LSL3: C-A-X now skips the questions
+ HQ: After "razzle dazzle root beer", A-X, A-K, A-T etc. work

Regarding the speed problems in Hoyle: There's one feature of the gfxop
toolset we have never /ever/ used so far, and that's explicit picviews.
This allows us to work the same way Sierra did; of course, we'll have
to keep track of the changes so that restoring a savegame will work.
This, however, only affects debugging savegames, since the games appear to
be able to restore the screen state themselves, so we might make it a
compile-time option and thus a development-only issue
(And a very minor one at that, since view widgets require something in
the order of 100 bytes of memory apiece).

BTW: I can't get xdos to run EGA games remotely on my home setup, so I'm
stuck without a way to test Sierra SCI for the time being. Could someone
please check the following in the original interpreter:
- Why does the dragon's flame in the HQ intro pass through the "Sierra
  presents" in Sierra SCI? The "Sierra presents" does not have IGNORE_ACTOR
  set, so the flame should be blocked by a collision (as in FreeSCI)
- Why and how does movement resume in the CB1 campus scene (also KQ4 intro
  regarding the fairies?) when the students' initial mover has finished?

Thanks in advance for any help!


-- Christoph