* Fixes and improvements to oldmidi.c by Rickard Lind * Fixed file name in sfx/ROADMAP (also by Rickard)
oldmidi.c is only used once, during startup (at least in the interpreter),
so the risk of breaking anything there is minimal.

Solomon: The ROADMAP fix changes the statement
"* change "oldmidi.c" to use the new device driver for MT-32"
to
"* change "soundserver_null.c" to use the new device driver for MT-32"
which appears to match nicely with your roadmap.


Regarding IRC: I'd like to apologize (also in advance) for mysteriously
disappearing from time to time. openprojects appears to be very stable,
but the line to my ISP isn't. It wasn't meant for a 24/7 connection, and
therefore my ISP appears to enjoy making a point out of closing my line at
appropriate times during the day.

-- Christoph