Okay, my first contribution in a while, albeit a rather minor one.
Fixed the "hang" while waiting for the pod to shut down in SQ1.
The game was setting the loop count of the sfx to "-1", which was being
interpreted as "loop forever".  I believe that this is really "play
once", as many one-off sfx in sq1 and other games show up with a loop
count of "-1".

More bug reports to follow.

 - Pizza