* xlib driver updates: Fixed braindead typo in line drawing * xlib driver: Added a few missing keys
Except for DGA support and the endianness issue, the xlib driver should be
complete now. Please test it and tell me what you think!

(Note that this driver allocates some memory; I'm not quite certain whether
xlib or the driver itself are supposed to handle that, so there may be a few
memory leaks.)

-- Christoph