* (a) Reverted yesterday's changes to parser tree augmentation * (b) AddToPic() now updates the dynview list's position * (c) DoBresen inceminates the b-stepCounter(?).
Observations:
= (a): See yesterday. Note that the meep accepts "ask for green green fur"
  instead.
+ (b) closes #83 (LSL3 voyeurism scene)
+ (c) closes #87 (first time I've seen the b-stepCount being used)

According to some tests, #73 and #75 are closed as well. LSL2 still can't
be finished because of #88.


BTW, I'm considering making the xlib driver the default if (getenv(DISPLAY)),
once the endianness stuff has been resolved and palette ops have been added.
Even without xshm, it's a lot faster. Any thoughts on that?

-- Christoph