* Re-wrote a single statement in the pic0 drawing code to something   semantically equivalent * Added missing 'break' statement in gfx_support.c (Thanks, Lars!)
The first one appears to have caused some major problems on ia32/Debian!
Some areas (QfG: Magic shop, inn) apparently were inaccessible
(or something like that). This worked fine on Solaris and even on Alpha,
so I guess this is a bug either in gcc or in Debian's patched version of
gcc (not neccessarily the most recent one).


-- Christoph