* Changed a few signal flag names * Re-implementation of Animate() according to more recent specs from Lars
Colission checking against dynviews is temporarily disabled because of this.
It doesn't appear to fix our visual problems (specifically #90, #70, #69, #56,
#26) yet (haven't checked all of them, though), but it should be closer to the
original algorithm. A quick comparison points out that flags are now almost
identical to the ones in Sierra SCI (with the exception of NO_UPDATE).

Please test it! This has the potential to both fix and break a wide variety of
stuff.

-- Christoph