Fixed some documentation.
This just in -- SQ3 is also broken -- when the scow picks up the pod in
the intro, the doors crack open, then suddenly are wide open.

The pod goes up, and the doors close to the crack again.. but the pod is
still visible.

That and the broken text in the intro.  it looks like it's not getting
dithered/blended in properly.  I think it was good before.