Added in reverb support to the mt32 driver.   Now sets the default reverb on initialization, and resets the reverb to that whenever it gets a new song.  (comments?  what games use reverb, anyway? just PQ2?)
Also re-added the rhythm channel map, which was broken in the original
source.  (dangling else problem..)

I guess I'm the only one with a mt-32.. but anyone with a midi board
should get the notes, even if the instruments are wrong.  Anyhow.  No
more work for now.  need a midi board to test stuff out.  :)

 - Pizza