Completely reworked the midiout layer; now supports multiple drivers ala the gfx subsystem.  Currently there is only the 'null' driver, but I'm going to port the "old" midiout_unixraw over next.
Got a bug though -- if you give the -g option a bogus parameter, it
ignores it and otherwise works.

for example, sciv -gfoo putters along happily instead of barfing with an
error.  Since I copied the same code to handle the midiout stuff, it too
is broken in the same way.

I'd like to have the midiout in the config file too, but my brain isn't
alert enough yet to try and grok config.l  :)

It's rather trivial, but anyhow.  I have other fish to fry at the
moment.  Like midiout_unixraw.  :)

 - Pizza