D'oh, I screwed up the last commitlog.
Basically, FreeSCI now has "official" midiout capabilities if you have a
MT32 and /dev/midi support.    That's hardwired, but as soon as I mangle
config.l even more, it'll be configurable.

So if you want to hear midi music, ./sciv -Munixraw

(and now for me to actually fire it up and listen to what breaks..)

Two other minor fixes -- removed Makefile.in from the cvs repository in
doc/man, and added a .cvsignore in there.

 - Pizza