* (a) Added check for ALSA (defines HAVE_ALSA) * (b) Simplified config.l
(a) is taken from alsa-utils-0.9.0beta1. Incidently, it was originally written
by Christopher Lansdown :-)

(b) It's quite a bit simpler now. Not quite as simple as it should be, but it
should be rather straightforward to use with the various OPT_XXX() macros, and
we're no longer building an FSM which requires extra options to build with
Solaris' lex.

Note that all of them require an _int_, except for OPT_STATICREF(), which
requires a pointer with the same size as sizeof(void *) (any pointer on the
machines we're targetting ATM). Note that this is an 'int' as in 'int', and
and not as in 'long', 'short', 'float', 'int32' or whatever.


Regarding (a) again, there are some problems with linking now, but I guess
those can be resolved by people who have a better idea of what should go
where.

Also note that I've upgraded my system and can no longer reproduce the freeze,
because the oss layers from ALSA 0.9.0 beta1 don't seem to work here.
(This means I'll have to wait for the ALSA implementation before I can comment
on anything related to audio output).

Incoming bugs:
#107: -g option accepts invalid values and falls back to default driver rather
than aborting
#108: gfxop_add_to_pic() destroys the original pic's dithering
#109: SQ3: Roger's eyes arent' removed when the intro finishes


Today's scores: 0:3, bugs win again
(I'll close the pod sound cue bug as soon as the other problems(?) regarding
the fix have been resolved)

-- Christoph