* (a) DoSound operation 0xb (TEST_DRIVER) now returns 32 on success rather than 1 * xlib driver returns mouse button pressed rather than 0 when checking button events * Applied Rickard's patch to improve the sound server * Included Petr's DirectDraw driver
(a) fixes the problem with no sound in Iceman, HQ, and QfG1. I assume that
later versions of SCI use this to return the driver's polyphony. This must
be made more dynamical to support PC speaker or Tandy output.
In case you're wondering, Sierra SCI appears to do the same. The result of
the calls to DoSound(TEST_DRIVER) are stored in global selector 0xd4 in
HQ, in case you want to check the values for other drivers.

Bug #96 has fixed itself.

BTW, note that I don't know how complete the DirectDraw driver is.

-- Christoph

VS: ----------------------------------------------------------------------