* (a) Fixed linear filter * Fixed menubar priorities * (b) Animate() opening animations now unclip the screen * (c) Alpha-blending crossblitter
Sorry for the lack of updates; (a) and (c) took quite a lot of my time
(and still aren't finished).

(a): The last filter was semi-broken, especially on high resolutions. This
one is better, but it still isn't what people traditionally refer to as
"bilinear". (I'll do that when, and if, I find the time to do so.)
(b): Fixed a rather noticeable problem in the SQ3 intro.
(c): Not everyone's going to be happy about this one... OK, here it comes:
In order to test whether the filter worked as intended (and because alpha
blending sounded like a cool idea), I wrote this little thing. It'll break
builds on Alpha systems right now, unless CC="gcc -mcpu=ev6" is used; this
is because it uses the pkwb and unpkbw operations, which are part of the
MVI instruction set extension.
This will work if and only if you're running the xlib target (I'll fix GGI
soon) on 32bpp on an EV6 or later (some of the late EV56 implementations
might work, too). I'm sorry for coding all of this mess in assembly, but I
have the following excuses:
a) It's just a feature
b) I haven't touched assembly in /years/ (and it's starting to show)
c) The code gcc produced when I wrote this in C looked... sub-optimal.
d) It's still slow as hell

The current implementation still needs to be updated to support 64 bit
read and write operations on the framebuffer (right now, those are only
done to the priority buffer).

I'd appreciate any feedback on this, BTW...


New bugs:
#116: Emmanuel Jeandel (IIRC) reported this one:
LSL3: Typing "look woman" at the beach doesn't work.

Probably missing synonyms.
I promise I'll go bug sqatting again RSN...


-- Christoph