Split the SET_VOLUME into SET_VOLUME and GET_VOLUME, much like the SET/GET MUTE is implemented.
It's cleaner this way...

one problem persists when restoring games.  If you restore a save game
that was saved with the sound muted, the volume gets messed up as well.
Is the main game loop (as in, SCI itself) setting the volume as part of
the restore?  (are my changes to the sfx_save code redundant?)

 - Pizza [and now, I really go to bed...]