The beginning of XShm support in the Xlib gfx driver.
Currently it keys off of the HAVE_MITSHM #define, and that's currently
disabled.  (in fact, all of the SHM code is in #ifdefs)

What this does:  Detects the XShm extension, tests it, and generally
sets stuff up for the rest of the code to use it.  There's no harm in
enabling the #define, as everything still works.

So, hows about adding XShm detection to the automake/autoconf scripts?
:)   You'll also need to link it against -lXext in order for it to work.

 - Pizza