Okay; here's a "I don't know if this works or not" commit of the XShm stuff.  It doesn't work for me, but I don't know if that's because my X server doesn't [properly] support Shared Pixmaps or not...
Basically, it tries to do the XShm Pixmap stuff, and if that fails, it
reverts back to regular Pixmaps with no hard feelings. (though you get a
couple of errors..)

Anyway, many eyeballs are better than one, so please, give this a shot
and let me know if it works for you.  It won't break anything simply by
being there... unless it works more for you than it does for me.

 - Pizza