Shared memory Xlib works now!
It's nice when you discover that all the docs are WRONG!  #$^@%&@%!!!

The main window isn't a shared pixmap yet; but I'll tackle that next.
All of the back buffers are though..

I'm interested to hear what kind of speed improvements this stuff makes
for everyone else... give it a shot, please!

...

Also -- A "new" bug:

QFG1:  The dragon's firey breath in the intro sticks near the top of the
screen; it looks like recent gfx changes broke it differently than it
was already broken.  :)

 - Pizza