here's the first commit of my work-in-progress SDL gfx target.
It won't compile (old xlib code I haven't replaced yet) and even if
it did, it won't work because it's lacking all of the buffer copying and
other minor things like that.

I'm committing it because I believe in data replication!

(for those who are wondering; the SDL code is #ifdef'ed out, so the cvs
 tree will continue to build cleanly)

 - Pizza