* (a) NO_UPDATE no longer implies that something cannot be collided against * caseopen() and fcaseopen() now bail out if a path separator character is present in the file name * (b) Fixes to functions using caseopen() and fcaseopen() * Added support for SDL 1.1.3
(a) partially fixes #172
(b) fixes problems reported regarding savegames (please test this!)

Closed bugs:
#169 Fixed yesterday
#134 (apparently no longer applies)
#162 Improved, but not gone- however, Xlib won't allow us to check for key
release events without holding the keyboard focus (and it's better that
way...), so this can't be fixed without changing the input model.

New bugs:
#170: Doors in LPD are off by one pixel (c)
#171: Door before shooting range doesn't open if approached from the outside
#172: Cop before evidence counter is drawn over Sonny (d)

(c) In which direction? Could somebody please check this and the priority,
coordinates, and rectangles of the cop in (d)? Remote DOSEmu dies on me
long before I get inside the PD building...


Bugs which should be re-checked:
#72, #87, #93, #100, #133

(all of these are potentially closed).

Also, are there any other FreeBSD users on this list? I tried #150 with a
clean FreeBSD-4.2 installation, and couldn't reproduce it.


-- Christoph