* Got rid of the default loop point on songs. * Loop only if the song set an explicit loop point.
This didn't fix the SQ3 pod bug (it sets the looppoint to 71) but it
should fix bug #180.  I can't verify this, as I get kernel oopses with
Hoyle2 still.

Also, can someone re-verify #151 and #153?  We're restoring the
insturment state when songs change back and forth, and in the case of
#151, that's been fixed too in the form of dumping the standard drum map
if it's not present in patch.001

 - Pizza