Initial checkin of the SDL threaded sound server.
Most of the code is still shared with the "null" soundserver, so I make
no guarantees that it will even compile if HAVE_FORK is not defined.

It is now the default soundserver if you have SDL installed.  Also,
this bumps the required version of SDL to 1.1.8.

Large parts of the old soundserver/sound code have been swapped around
and refactored.   So there's the potential here for much breakage!

Oh, we should probably remove the completely obselete soundserver_dos.c
now.

 - Pizza